Rewrite SysEx parser, add SysEx writer
This rewrites the incorrect SysEx parser and adds a SysEx writer.
Informal testing has shown that parsing and then writing produces a
byte-for-byte identical copy in the majority of patches (the writer
will not reproduce invalid patches because the parser clamps values
to their valid ranges). This also adds oscillator key syncing to
voices, as this was missing.